+++ frontend/lost/src/components/SIA/lost-sia/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0]
 ### Fixed 
 - Added camera move on wasd keys
+- Provide canvas key events via prop callback
 
 ## [0.1.1] - 2020-04-06
 ### Fixed